Refactor file stream methods and fix GetFileSizeAsync implementation
Updated the streaming methods in ProviderFileDao, GoogleCloudStorage, S3Storage and FileHandler to replace usage of int.MaxValue with long.MaxValue, thus supporting larger file sizes. Additionally, the GetFileSizeAsync method for SharePointFileDao has been fixed to provide actual file size instead of returning a default value.

+++ common/ASC.Data.Storage/GoogleCloud/GoogleCloudStorage.cs
@@ -145,7 +145,7 @@ public class GoogleCloudStorage(TempStream tempStream,
 
     public override async Task<Stream> GetReadStreamAsync(string domain, string path, long offset)
     {
-        return await GetReadStreamAsync(domain, path, offset, int.MaxValue);
+        return await GetReadStreamAsync(domain, path, offset, long.MaxValue);
     }
     
     public override async Task<Stream> GetReadStreamAsync(string domain, string path, long offset, long length)
@@ -155,7 +155,7 @@ public class GoogleCloudStorage(TempStream tempStream,
         var storage = await GetStorageAsync();
         DownloadObjectOptions options = null;
         
-        if (length > 0 && (offset > 0 || offset == 0 && length != int.MaxValue))
+        if (length > 0 && (offset > 0 || offset == 0 && length != long.MaxValue))
         {
             options = new DownloadObjectOptions
             {
